refactor(chatbot): organize session management into models and services with interfaces
